REC: fix MSVC warnings
git-svn-id: https://svn.osgeo.org/gdal/trunk@31610 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/rec/ogrreclayer.cpp
@@ -197,7 +197,7 @@ OGRFeature * OGRRECLayer::GetNextUnfilteredFeature()
         }
 
         // If the end-of-line markers is '?' the record is deleted.
-        iSegLen = strlen(pszLine);
+        iSegLen = (int)strlen(pszLine);
         if( pszLine[iSegLen-1] == '?' )
         {
             pszRecord[0] = '\0';
